1-py-test/old-framework-test/requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
huggingface-hub>=0.16.4
tqdm>=4.65.0
logging>=0.4.9.6
pyyaml>=6.0.1
numpy>=1.24.3
matplotlib>=3.9.0